serverless/requirements_serverless.txt
azure-eventhub==5.11.5
azure-functions==1.17.0
